--- a/aiida/backends/sqlalchemy/models/authinfo.py
+++ b/aiida/backends/sqlalchemy/models/authinfo.py
@@ -55,5 +55,5 @@ class DbAuthInfo(Base):
 
     def get_aiida_class(self):
         from aiida.orm.implementation.sqlalchemy.authinfo import SqlaAuthInfo
-        from aiida.orm.backends import construct_backend
-        return SqlaAuthInfo.from_dbmodel(dbmodel=self, backend=construct_backend())
+        from aiida.manage import get_manager
+        return SqlaAuthInfo.from_dbmodel(dbmodel=self, backend=get_manager().get_backend())