--- a/aiida/backends/sqlalchemy/models/user.py
+++ b/aiida/backends/sqlalchemy/models/user.py
@@ -50,5 +50,5 @@ class DbUser(Base):
 
     def get_aiida_class(self):
         from aiida.orm.implementation.sqlalchemy.users import SqlaUser
-        from aiida.orm.backends import construct_backend
-        return SqlaUser.from_dbmodel(self, construct_backend())
+        from aiida.manage import get_manager
+        return SqlaUser.from_dbmodel(self, get_manager().get_backend())