--- a/aiida/cmdline/commands/cmd_data/cmd_bands.py
+++ b/aiida/cmdline/commands/cmd_data/cmd_bands.py
@@ -46,11 +46,11 @@ def bands():
 @options.FORMULA_MODE()
 def bands_list(elements, elements_exclusive, raw, formula_mode, past_days, groups, all_users):
     """List BandsData objects."""
-    from aiida.orm import construct_backend
+    from aiida.manage import get_manager
     from tabulate import tabulate
     from argparse import Namespace
 
-    backend = construct_backend()
+    backend = get_manager().get_backend()
 
     args = Namespace()
     args.element = elements