--- a/aiida/cmdline/commands/cmd_process.py
+++ b/aiida/cmdline/commands/cmd_process.py
@@ -18,7 +18,7 @@ from aiida.cmdline.params import arguments, options
 from aiida.cmdline.utils import decorators, echo
 from aiida.cmdline.utils.query.calculation import CalculationQueryBuilder
 from aiida.common.log import LOG_LEVELS
-from aiida.manage.manager import AiiDAManager
+from aiida.manage import get_manager
 
 
 @verdi.group('process')
@@ -120,7 +120,7 @@ def process_status(processes):
 def process_kill(processes, timeout, wait):
     """Kill running processes."""
 
-    controller = AiiDAManager.get_process_controller()
+    controller = get_manager().get_process_controller()
 
     futures = {}
     for process in processes:
@@ -147,7 +147,7 @@ def process_kill(processes, timeout, wait):
 def process_pause(processes, timeout, wait):
     """Pause running processes."""
 
-    controller = AiiDAManager.get_process_controller()
+    controller = get_manager().get_process_controller()
 
     futures = {}
     for process in processes:
@@ -174,7 +174,7 @@ def process_pause(processes, timeout, wait):
 def process_play(processes, timeout, wait):
     """Play paused processes."""
 
-    controller = AiiDAManager.get_process_controller()
+    controller = get_manager().get_process_controller()
 
     futures = {}
     for process in processes:
@@ -201,7 +201,7 @@ def process_watch(processes):
     def _print(body, sender, subject, correlation_id):
         echo.echo('pk={}, subject={}, body={}, correlation_id={}'.format(sender, subject, body, correlation_id))
 
-    communicator = AiiDAManager.get_communicator()
+    communicator = get_manager().get_communicator()
 
     for process in processes: