--- a/aiida/common/__init__.py
+++ b/aiida/common/__init__.py
@@ -7,7 +7,15 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
+"""AiiDA common functionality"""
+
 from __future__ import division
 from __future__ import print_function
 from __future__ import absolute_import
 from aiida.common.log import AIIDA_LOGGER
+
+# pylint: disable=wildcard-import
+
+from .profile import *
+
+__all__ = profile.__all__