* move AiiDAManager from `aiida.work` to `aiida.manage`
 * introduce caching of .objects in `aiida.orm.entities.Collection`
 * cache default user in `aiida.orm.user.User.Collection` (fix #2216)

## Benchmark on MacBook Pro, python3
  * before: 58.5s to create 10k nodes (57.2 spent inside `Node.__init__`)
  * after: 2.8s to create 10k nodes (1.7s spent inside `Node.__init__`)


The timing of 58.5s is significantly worse than what we wrote down during the tests with @szoupanos  and @giovannipizzi  (which was ~10s for 10k nodes), i.e. probably someone broke the caching.

Anyhow, the target we set then was: 2s for 10k nodes, which we have reached now on python3 (it's the 1.7s that should scale linearly with the number of nodes).