@ltalirz 2018-11-18
For performance tests, I opened #2215 

This PR is still a little bit dirty and I have a few questions, perhaps also for @sphuber or @muhrin :

 1. What exactly should be the scope of the `AiiDAManager`? Before this PR it looked backend-independent on the surface (I might be wrong) but now we are definitely adding backend-dependent stuff.
 1. @muhrin's work was making it possible to specify the backend for getting objects. I guess we should retain this functionality in the AiiDA manager. To me this means we either
    a. cache the backend in the AiiDA manager and use that and/or 
    b. make it possible to specify the backend when accessing the cache (and effectively have a per-backend cache)

 1. My purpose here was just to speed up `Node()` but the uncached `User.objects.get_default()` is used in many other places in the code. We'll need to decide on a policy of where to use cached objects and where not.