@muhrin 2018-11-18
Dear Leo, thanks for the work - looks good.

My response:

1. `AiiDAManager` is specifically act as the single place to construct/cache/access backend specific calsses.  Particularly some which are quite complex to construct e.g. `Runner`s and `Communicator`s.  So I'm not sure what exactly you mean by 'backend dependent' but if you mean it does different things based on the current backend (or more specifically profile) then that's exactly what it's there for. 

2.  This is a good point and needs a little discussion.  For now AiiDAManager is a singleton of sorts (given that it only has class methods) that uses the current profile to make decisions but that precludes the ability to use it with different profiles simultanously...that said, this is kind of the point in that it's there to do the right thing based on the currently set profile (which could potentially be changed at runtime).

3. I would make this lazily cached in the `User` collection so it's fast for everyone using it