@ltalirz 2018-11-18
@muhrin thanks a lot for the thoughts!

Re 2.: Ok. So, the idea is that when you load a new profile (when this becomes possible...), the AiiDAManager  will be reset, correct? I will add this to its docstring.

Re 3.: Could you elaborate on this a bit? In 1. you wrote that the `AiiDAManager` should be the single place to construct/cache/access backend specific classes.
So with`lazily cached` do you mean:
  * `The default user will be cached inside the `AiiDAManager` 
  * `User.objects.get_default()` will call `AiiDAManager.get_default_user()`
Furthermore, is there currently anything that invalidates the cache?
What if you pass to `User.objects.get_default()` a backend that is not the current one?