@muhrin 2018-11-18
2. Correct.

3. The dependency relationship was envisaged to be that `AiiDAManager` know about the backend but not the other way around (where possible).  This would mean that `AiiDAManager` should eventually contain the `construct_backend()` method and I would imagine that it would be `User.objects.get_default()` that would contain the caching itself.  If you do `User.objects(my_backend).get_default()` the first part (`User.objects(my_backend)`) actually passes back a separate collection instance which would itself contain the same caching logic as the default one.  Currently calling with a custom backend will not result in the collection being cached after instantiation which would have serious performance implications so it's a target for future optimisation.