@ltalirz 2018-11-18
> Currently calling with a custom backend will not result in the collection being cached after instantiation which would have serious performance implications so it's a target for future optimisation.

This holds even when calling with the default backend, correct? I.e. caching stuff inside `User.objects` currently does not work.

So, for the moment, should we cache the default user inside the `AiiDAManager` and use this cache from `User.objects.get_default()` with a notice to move this cache into `User.objects`  (and knowing that it will not do the right thing when passing a non-default backend to User.objects)?