@muhrin 2018-11-19
Hmm...I thought it was there already but you're right, it's not.  The place for caching (the default) objects collection is here:

https://github.com/aiidateam/aiida_core/blob/aa7907260906f01d8d7bb04b1b1421b24ab00d3f/aiida/orm/entities.py#L116

It's quick to add, happy to do it because the performance implications could be great.

If you get this PR through I can add the caching, you can track it here: https://github.com/aiidateam/aiida_core/issues/2216

I would avoid putting the caching in the `AiiDAManager` because 1) it will move anyway, and 2) I haven't really reflected on if this is even a good place to put things like this (although on the surface I don't see why not).