@muhrin 2018-11-29
Ok, ready to go.  @ltalirz please have a look. @sphuber please review