@ltalirz 2018-11-29
@muhrin added a few comments, otherwise looks good to me!
just the failing tests to fix...